+++ apps/web/src/app/AppLayoutWrapper.tsx
@@ -8,7 +8,7 @@ import {
 	WorkspaceSwitcher,
 } from "@/components/Modals";
 import SearchCommand from "@/components/SearchCommand";
-import { Toaster } from "@/components/ui/toaster";
+import { Toaster } from "@/components/ui/sonner";
 import { SquaredStoreProvider } from "@/store";
 import { ClerkProvider } from "@clerk/nextjs";
 import {