+++ packages/fonts/src/index.ts
@@ -1,3 +1,0 @@
-import "./styles.css";
-
-export { default as noirdenSansPlugin } from "../tailwind-plugin";